requirements.txt
Flask
langchain
openai
python-dotenv